+++ chat-widget/src/contexts/createReducer.ts
@@ -767,6 +767,23 @@ const reducer = (state: ILiveChatWidgetContext, action: ILiveChatWidgetAction): 
                     isConversationalSurvey: action.payload as boolean
                 }
             };
+
+        case LiveChatWidgetActionType.SET_USER_AUTHENTICATED:
+            inMemory = {
+                ...inMemory,
+                appStates: {
+                    ...inMemory.appStates,
+                    hasUserAuthenticated: action.payload as boolean
+                }
+            };
+            return {
+                ...state,
+                appStates: {
+                    ...state.appStates,
+                    hasUserAuthenticated: action.payload as boolean
+                }
+            };
+
         case LiveChatWidgetActionType.GET_IN_MEMORY_STATE:
             return inMemory;
         default: